requirements.txt
asttokens==2.4.1
black==24.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
executing==2.0.1
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
load-dotenv==0.1.0
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.3
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pyzmq==25.1.2
six==1.16.0
stack-data==0.6.3
tabula-py==2.9.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
tzdata==2024.1
wcwidth==0.2.13
